All replacement edit sections in ResumeEditor have been integrated
